+++ src/app.py
@@ -6,7 +6,7 @@ from telegram.ext.dispatcher import Dispatcher
 from logger import Logger
 from config.credentials import BOT_TOKEN, PORT, WEBHOOK_URL
 from bot.handler import (groupsHandler, joinHandler, mentionHandler, leaveHandler,
-                     silentMentionHandler, startHandler)
+                     silentMentionHandler, startHandler, inlineQueryHandler)
 from bot.handler.abstractHandler import AbstractHandler