+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
  
+## [UNRELEASED] - 07.10.2021
+### Added
+- Inline Query for join/leave/everyone
+- Validator class for group name
+
+### Updated
+- start command content
 ## [0.1.0] - 06.10.2021
 ### Features
 - `/join` command